Add explore functionality
inventory is the first implementation for the collector. It adds a cli tool
to explore json files and show their content. In addition, it has provides
create databases for infrastructures without "PBS/Torque".

In addition an extension to the collector has been implemented, that collect
additional informations like network configuration, active users, mount points
and running or inactive vms.

The database class is the connection to the json file, an additional
layer which can be used to interact with json files.

Change-Id: I72ea1378e7a00bcd13fc98c06aa799e2c2b6cad3
Signed-off-by: Uwe Schilling <uwe.schilling@hlrs.de>